refactor(triage): split triage.rs god-file into modules
