+++ src/core/triage/gh.rs
@@ -1,0 +1,62 @@
+//! gh CLI invocation helpers and shared GitHub status-check summarization.
+
+use std::process::Command;
+
+use serde_json::Value;
+
+use crate::core::git::gh_probe_succeeds;
+
+pub(crate) fn ensure_gh_ready() -> std::result::Result<(), String> {
+    if !gh_probe_succeeds(&["--version"]) {
+        return Err("gh CLI not found on PATH".to_string());
+    }
+    if !gh_probe_succeeds(&["auth", "status", "--hostname", "github.com"]) {
+        return Err("gh is not authenticated for github.com".to_string());
+    }
+    Ok(())
+}
+
+pub(crate) fn run_gh(args: &[String]) -> std::result::Result<String, String> {
+    let output = Command::new("gh")
+        .args(args.iter().map(|s| s.as_str()))
+        .output()
+        .map_err(|e| format!("failed to invoke gh: {e}"))?;
+    if !output.status.success() {
+        let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
+        let stdout = String::from_utf8_lossy(&output.stdout).trim().to_string();
+        return Err(if stderr.is_empty() { stdout } else { stderr });
+    }
+    Ok(String::from_utf8_lossy(&output.stdout).to_string())
+}
+
+pub(crate) fn non_empty(value: Option<String>) -> Option<String> {
+    value.and_then(|v| {
+        let trimmed = v.trim();
+        if trimmed.is_empty() {
+            None
+        } else {
+            Some(trimmed.to_string())
+        }
+    })
+}
+
+pub(crate) fn summarize_checks(checks: &[Value]) -> Option<String> {
+    if checks.is_empty() {
+        return None;
+    }
+    let mut saw_pending = false;
+    for check in checks {
+        let conclusion = check.get("conclusion").and_then(Value::as_str);
+        let status = check.get("status").and_then(Value::as_str);
+        if matches!(
+            conclusion,
+            Some("FAILURE" | "CANCELLED" | "TIMED_OUT" | "ACTION_REQUIRED")
+        ) {
+            return Some("FAILURE".to_string());
+        }
+        if conclusion.is_none() && !matches!(status, Some("COMPLETED")) {
+            saw_pending = true;
+        }
+    }
+    Some(if saw_pending { "PENDING" } else { "SUCCESS" }.to_string())
+}
